Create MovieAPI

+++ src/main/java/com/sda/awesomemovies/api/movie/MovieRepository.java
@@ -2,6 +2,6 @@ package com.sda.awesomemovies.api.movie;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface MovieRepository extends CrudRepository<MovieEntity, Long> {
+public interface MovieRepository extends CrudRepository<MovieEntity, Integer> {
     MovieEntity findByTitle(String title);
 }
